fix(pptx,core): correct bar-chart data-label colour, comma, and visibility
Sample-2 slide-16's stacked bar chart rendered its data labels wrong on
three counts vs PowerPoint:

1. Colour — the chart sets per-series label colours via
   `<c:ser><c:dLbls><c:txPr>` (white "bg1" on the dark segment, black "tx1"
   on the light one, ECMA-376 §21.2.2.216). The parser only extracted a
   single chart-level `dataLabelFontColor`, so the light segment's "7,507"
   rendered white and illegible. Add a per-series `labelColor` and prefer it
   over the chart-level colour in the bar renderer.

2. Thousands separator — the series number format
   `#,##0_);[Red]\(#,##0\)` was never read from
   `<c:val><c:numCache><c:formatCode>` (§21.2.2.121), so labels showed
   "7507" instead of "7,507". Parse it into `valFormatCode`; the renderer
   already prefers it when `<c:dLbls>` has no explicit `<c:numFmt>`.

3. Centering — already correct (dLblPos=ctr); the value only looked
   misaligned because it was invisible (white-on-light) before fix #1.

Verified against the PowerPoint-exported PDF (sample-2 page 16 + the charts
on page 7): labels now match colour, comma formatting, and the
stacked/centred layout.

Also extract the chart value-formatting helpers into a pure, unit-tested
`chart-number-format.ts` (same pattern as `axis-scale.ts`), locking the
§18.8.30 section/thousands behaviour. Rendering is byte-identical (0-pixel
diff on slide-16 before/after the extraction).
